Add Book constructor and prototype, myLibrary arr
To store basic information about a book, along with the user's read
status of the book, a Book constructor was added to scripts.js.

To obtain the user's read status of a book, change its read status, and
return a string of information regarding the book, the Book prototype
was modified to include methods that perform the aforementioned actions
(getReadStatus, toggleReadStatus, getBookInfo).

For storing the user's books (Book objects), the myLibrary array was
added to scripts.js.