Add 'extractFormData', 'processNewBook' functions
To extract all of the values from the 'add a book' form and return them
in an array, the 'extractFormData' function was added to scripts.js.

To create a new book and add it to the library based on the values that
were submitted from the 'add a book' form, the 'processNewBook' function
was added to scripts.js.