Add 'createBookButtonsDiv' fn to 'CardElement'
For each CardElement to include a section for buttons that provides
options to the user for manipulating a selected book, the CardElement
constructor was modified to include a 'createBookButtonsDiv' function.

To provide the option for a user to delete a selected book, the
'createDeleteButton' function was added within the
'createBookButtonsDiv' function.