Add 'removeCard', 'deleteObjectKeys' functions
For the 'card' HTML structure to be removed from the page after the user
has clicked the 'Delete' button on a selected book, as well as remove
the object containing the references to the 'card' HTML structure from
the 'cardElementArr' array, the 'removeCard' function was added.

To ensure that the memory containing the references to the elements
for the removed 'card' HTML structure does not persist, the
'deleteObjectKeys' function was added.