Add 'reindexCardElementArr' function
Once a book and the 'card' HTML structure has been deleted, the
'data-index' of the other books on the HTML pages have not been updated
to take into account the deleted book, resulting in issues performing
actions on the current books, as well any books that were to be added
afterwards.

To re-index the 'data-index' of current books on the HTML page, the
'reindexCardElementArr' function was added.